unify summarization logic
